scrapmon/requirements.txt
django-celery
redis